add: Componente ProfileCard
